refactor(LottoException): 패키지명 수정
LottoException의 패키지명을 constants로 수정하였습니다.
