+++ README.md
@@ -9,6 +9,7 @@
     
 2. 로또 발행
     - [ ] 구입한 로또의 수 만큼 랜덤 번호 발행
+    - [ ] 구입한 로또와 당첨 번호 비교
     - [ ] 당첨 여부 기록
     - [ ] 수익률 계산
 
@@ -28,6 +29,7 @@
 2. 당첨 번호 입력
     - [X] 구분자(,)로 split했을 때 문자가 포함된 경우
     - [X] 숫자가 6개가 아닌 경우
+    - [X] 중복된 값이 입력된 경우
     - [X] 로또 번호 범위를 벗어나는 경우
 
 
@@ -39,7 +41,6 @@
 ```text
 ├── model/
 │   ├── Lotto.java                    # 로또 한 장
-│   ├── WinningLotto.java             # 당첨 번호 정보
 │   ├── LottoGenerator.java           # 로또 번호 생성
 │   ├── LottoMatcher.java             # 당첨 번호 매칭 
 │   ├── LottoResult.java              # 당첨 결과 정보
@@ -51,7 +52,7 @@
 │   └── OutputView.java               # 모든 출력을 처리
 ├── validator/
 │   └── LottoValidator.java           # 입력값 검증
-└── exception/
+└── constants/
     └── LottoException.java           # 예외 정의 (enum)
 ```
