+++ src/main/java/validator/LottoValidator.java
@@ -5,7 +5,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-import static exception.LottoException.*;
+import static constants.LottoException.*;
 
 public class LottoValidator {
     public static Integer isNumber(String targetString) {
